+++ HouseholdAccountBook/app/src/main/java/myclasses/IncomeCategory.java
@@ -1,7 +1,7 @@
 package myclasses;
 
 public class IncomeCategory extends BopCategory {
-    public IncomeCategory(Integer id, String name, int colorCode, int index, boolean isDeleted) {
+    public IncomeCategory(Long id, String name, int colorCode, int index, boolean isDeleted) {
         super(id, name, colorCode, index, isDeleted);
     }
 }
